pyproject.toml
[project]
name = "pixelpose"
version = "0.1.0"
description = "A dataset of pixelated characters with poses"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=1.24.0",
    "opencv-python>=4.12.0.88",
    "Pillow>=10.0.0",
    "tqdm>=4.67.1",
    "requests>=2.31.0",
    "yaspin>=3.2.0",
    "ipykernel>=6.30.1",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "diffusers>=0.35.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "accelerate>=1.10.1",
    "open-clip-torch>=3.2.0",
    "ipywidgets>=8.1.7",
    "transformers>=4.57.0",
    "pytest>=8.4.2",
    "wandb>=0.22.2",
    "click>=8.3.0",
    "tenacity>=9.1.2",
]

[project.scripts]
pixelpose = "pixelpose.app:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []